src/mb: Fix Display render thread; remove clutter
* src/mb/Display.cpp: Likewise.
* src/mb/Display.h: Likewise.
* src/ctxmgmt/SystemInfo.cpp: Updated accordingly.
* src/formatter/FormatterActiveDevice.cpp: Likewise.
* src/formatter/FormatterMultiDevice.cpp: Likewise.
* src/formatter/NclFormatterRegion.cpp: Likewise.
* src/lssm/PresentationEngineManager.cpp: Likewise.
* src/mb/SDLAudioProvider.cpp: Likewise.
* src/mb/SDLAudioProvider.h: Likewise.
* src/mb/SDLVideoProvider.cpp: Likewise.
* src/mb/SDLVideoProvider.h: Likewise.
* src/mb/SDLWindow.cpp: Likewise.
* src/mb/SDLWindow.h: Likewise.
* src/player/AVPlayer.cpp: Likewise.
* src/player/ShowButton.cpp: Likewise.
* tests-ncl/test-image-png-mirror.ncl: Fix src.
* contrib/tests-env.sh: Abort on errors only.

+++ src/formatter/NclFormatterRegion.cpp
@@ -777,7 +777,7 @@ NclFormatterRegion::prepareOutputDisplay (SDLSurface* renderedSurface,
       if (!externHandler)
         {
           outputDisplay = Ginga_Display
-            ->createWindow (left, top, width, height, cvtIndex);
+            ->createWindow (left, top, width, height, (int)cvtIndex);
         }
 
       clog << "NclFormatterRegion::prepareOutputDisplay '" << outputDisplay;
